requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-oauth-toolkit==1.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
pep8==1.7.1
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
